Refactor jvm stdlib with code refactor
