+++ stdlib/jvm/src/main/java/org/ballerinalang/nativeimpl/jvm/methodvisitor/VisitMethodInsn.java
@@ -38,8 +38,7 @@ import static org.ballerinalang.nativeimpl.jvm.ASMUtil.METHOD_VISITOR;
 @BallerinaFunction(
         orgName = "ballerina", packageName = "jvm",
         functionName = "visitMethodInsn",
-        receiver = @Receiver(type = OBJECT, structType = METHOD_VISITOR,
-                structPackage = JVM_PKG_PATH),
+        receiver = @Receiver(type = OBJECT, structType = METHOD_VISITOR, structPackage = JVM_PKG_PATH),
         args = {
                 @Argument(name = "opcode", type = INT),
                 @Argument(name = "className", type = STRING),
